LayoutTests/fast/loader/disallow-redirect-to-invalid-url-using-meta-refresh-expected.txt

frame "frame" - didStartProvisionalLoadForFrame
main frame - didFinishDocumentLoadForFrame
frame "frame" - didCommitLoadForFrame
frame "frame" - didFinishDocumentLoadForFrame
frame "frame" - didHandleOnloadEventsForFrame
main frame - didHandleOnloadEventsForFrame
frame "frame" - didFinishLoadForFrame
main frame - didFinishLoadForFrame
Tests that we do not redirect to an invalid URL initiated by <meta http-equiv="refresh"> when setting allowNavigationToInvalidURL is false. This test PASSED if you do not see an entry in the dumped frame load callbacks of the form: "willPerformClientRedirectToURL: x-dummy://A=a&B=b".

Note, this test must be run in DumpRenderTree.
